mcp/requirements.txt
mcp
fastmcp
chromadb
sentence-transformers
pyyaml
